Cleaning and refractorating code
Made a few new methods in Shop to allow me to see if product exist and find index of product. The indexfinding method is used to return product to add to Customer's cart. Trying to make the program as loose as possible which i hope I've managed to. Feedback is very appricieded.