Further codestructurering + ID control start
Developed further on cleaning up code and removing wasted code. MAde the structure a lot better as well added ID to Account. It is generated in the class itself without any constructor input. In Account, it takes the method from miscUtility and uses it in a private method called generateID. It basically calls from the miscUtility.