Fixed previous problem
Had a problem with employee and employer accessing to both Employee and Employer caused by they are sharing the same AccountType when creating the ArrayList. Fixed by having a protected AccountType in Account (also added this to customerAccount if I ever wanted to genericify the Account arraylist. Gets the accountType by the method getAccountType which is in Account which is a super class.